@@ -1,3 +1,3 @@
 class SubmissionViewLog < ActiveRecord::Base
-  attr_accessible :submission_id, :user_id
+  #attr_accessible :submission_id, :user_id
 end
