@@ -1,4 +1,4 @@
 class Testcase < ActiveRecord::Base
   belongs_to :problem
-  attr_accessible :group, :input, :num, :score, :sol
+  #attr_accessible :group, :input, :num, :score, :sol
 end
