@@ -6,8 +6,8 @@ CafeGrader::Application.configure do
   # since you don't have to restart the web server when you make code changes.
   config.cache_classes = false
 
-  # Log error messages when you accidentally call methods on nil.
-  config.whiny_nils = true
+  # Log error messages when you accidentally call methods on nil.  //DEPRICATED
+  # config.whiny_nils = true // DEPRICATED
 
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
@@ -23,11 +23,11 @@ CafeGrader::Application.configure do
   config.action_dispatch.best_standards_support = :builtin
 
   # Raise exception on mass assignment protection for Active Record models
-  config.active_record.mass_assignment_sanitizer = :strict
+  # config.active_record.mass_assignment_sanitizer = :strict //DEPRICATED
 
-  # Log the query plan for queries taking more than this (works
-  # with SQLite, MySQL, and PostgreSQL)
-  config.active_record.auto_explain_threshold_in_seconds = 0.5
+  # Log the query plan for queries taking more than this (works // DEPRICATED
+  # with SQLite, MySQL, and PostgreSQL) // DEPRICATED
+  # config.active_record.auto_explain_threshold_in_seconds = 0.5 // DEPRICATED
 
   # Do not compress assets
   config.assets.compress = false
@@ -36,5 +36,7 @@ CafeGrader::Application.configure do
   config.assets.debug = true
 
   # Prevents assets from rendering twice
-  config.serve_static_assets = true
+  config.serve_static_files = true
+
+  config.eager_load = false
 end
